Added verification to check if the player already started the quest.
